==> notebooks/01_data_cleaning.ipynb <==
CHARGEMENT DES LIBRAIRIES ET LE DATASET

import pandas as pd
import numpy as np
import os

# Pour visualiser rapidement les infos du dataset
pd.set_option('display.max_columns', None)
pd.set_option('display.max_rows', 20)

# Versions (reproductibilité)
print("pandas version:", pd.__version__)
print("numpy version:", np.__version__)

# Charger le dataset
df = pd.read_excel("../data/raw/OnlineRetail.xlsx")

# Afficher les premières lignes
df.head(6)

COMPRÉHENSION DU DATASET

COMPRENDRE LES DONNÉES

# Informations générales
df.info()

# Statistiques descriptives
df.describe()

# Vérification des doublons
print("Nombre de doublons initiaux :", df.duplicated().sum())

# Valeurs manquantes
df.isnull().sum()

# Colonnes initiales
print("Colonnes brutes :", df.columns.tolist())


STANDARDISATION DES NOMS DE COLONNES

NETTOYER LES DONNÉES

# Standardisation des noms de colonnes pour cohérence du projet
# Standardisation des noms de colonnes pour le projet
df.rename(columns={
    'Invoice': 'InvoiceNo',
    'Price': 'UnitPrice',
    'Customer ID': 'Customer_ID'
}, inplace=True)

print("Colonnes après renommage :", df.columns.tolist())


NETTOYAGE DES DONNÉES

# Suppression des factures annulées (InvoiceNo commençant par 'C')
df = df[~df['InvoiceNo'].astype(str).str.startswith('C')]

# Suppression des lignes sans Customer_ID
df = df.dropna(subset=['Customer_ID'])

# Conversion du type
df['Customer_ID'] = df['Customer_ID'].astype(int)

# Correction du type de Date
df['InvoiceDate'] = pd.to_datetime(df['InvoiceDate'])

GESTION DES DOUBLONS

# Suppression des doublons exacts
df = df.drop_duplicates()

print("Nombre de lignes après suppression des doublons :", len(df))

CRÉATION DE LA MÉTRIQUE BUSINESS CLÉ: REVENUE

VÉRIFIER LES ANOMALIES

# Calcul du chiffre d'affaires par ligne
df['Revenue'] = df['Quantity'] * df['UnitPrice']

TRAITEMENT DES VALEURS INCOHÉRENTS

# Détection des revenus négatifs ou nuls
neg_revenue = df[df['Revenue'] <= 0]
print("Nombre de ventes négatives ou nulles :", len(neg_revenue))

# Suppression pour analyses business
df = df[df['Revenue'] > 0]

EXPORT DES DONNÉES NETTOYÉES

EXPORTER LES DONNÉES NETTOYÉES

# Chemin racine du projet
project_root = os.path.abspath(os.path.join(os.getcwd(), ".."))

# Création du dossier processed si nécessaire
processed_path = os.path.join(project_root, "data", "processed")
os.makedirs(processed_path, exist_ok=True)

# Sauvegarde
csv_file_path = os.path.join(processed_path, "online_retail_clean.csv")
df.to_csv(csv_file_path, index=False)

print(f"Fichier nettoyé sauvegardé dans : {csv_file_path}")
