==> notebooks/02_eda.ipynb <==
IMPORT ET CHARGEMENT DES DONNÉES NETTOYÉES

import pandas as pd
import numpy as np
import plotly.express as px
import plotly.graph_objects as go

pd.set_option('display.max_columns', None)

# Chargement des données nettoyées
df = pd.read_csv("../data/processed/online_retail_clean.csv")
df.head(6)

KPI GLOBAUX

# Calcul des métriques globales du business
total_revenue = df['Revenue'].sum()           # Chiffre d'affaires total
total_customers = df['Customer_ID'].nunique() # Nombre de clients uniques
total_orders = df['InvoiceNo'].nunique()      # Nombre de commandes uniques
total_products = df['StockCode'].nunique()    # Nombre de produits différents vendus

total_revenue, total_customers, total_orders, total_products

EVOLUTION DU CHiFFRE DANS LE TEMPS

# Conversion de la date en format datetime et création d'une colonne mois-année
df['InvoiceDate'] = pd.to_datetime(df['InvoiceDate'])
df['YearMonth'] = df['InvoiceDate'].dt.to_period('M').astype(str)

# Agrégation du chiffre d'affaires par mois
Monthly_revenue = (
    df.groupby('YearMonth')['Revenue']
    .sum()
    .reset_index()
)

# Visualisation de l'évolution du chiffre d'affaires mensuel avec Plotly
fig = px.line(
    Monthly_revenue,    # DataFrame contenant les données agrégées
    x='YearMonth',      # Axe X : périodes mensuelles
    y='Revenue',        # Axe Y : chiffre d'affaires
    title='Monthly Revenue Trend',  # Titre du graphique
    markers=True        # Affiche un point pour chaque mois
)

# Personnalisation de la mise en page
fig.update_layout(
    xaxis_title = "Month",          # Libellé axe X
    yaxis_title = "Revenue",        # Libellé axe Y
    template = "plotly_white"       # Thème visuel épuré (fond blanc)
)

fig.write_image("../reports/figures/monthly_revenue_Trend.png")
# Affichage interactif du graphique
fig.show()

CHIFFRE D'AFFAIRES PAR PAYS (PRIORISATION MARCHÉS)

# Agrégation du chiffre d'affaires par pays, trié du plus élevé au plus faible
country_revenue = (
    df.groupby('Country')['Revenue']  # Groupement par pays
    .sum()                            # Somme des revenus pour chaque pays
    .sort_values(ascending=False)     # Tri décroissant (meilleurs pays en premier)
    .reset_index()                    # Réinitialise l'index pour obtenir un DataFrame propre
)

# Création d'un graphique en barres des 10 pays les plus rentables
fig = px.bar(
    country_revenue.head(10),    # Sélection des 10 premiers pays (CA le plus élevé)
    x='Country',                 # Axe X : noms des pays
    y='Revenue',                 # Axe Y : chiffre d'affaires
    title='Top 10 Countries by Revenue',  # Titre du graphique
    text_auto='.2s'              # Affiche les valeurs sur les barres (format abrégé : 1.2k, 1.2M...)
)

# Personnalisation de la mise en page
fig.update_layout(
    template="plotly_white",              # Thème visuel épuré avec fond blanc
    yaxis={'categoryorder':'total ascending'}  # Trie les barres par ordre croissant de valeur
)

# Affichage interactif du graphique
fig.write_image("../reports/figures/Countries_revenue.png")
fig.show()

CLIENTS A HAUTE VALEUR (FONDATION DU RFM)

# Agrégation du chiffre d'affaires par client, trié du plus élevé au plus faible
customer_revenue = (
    df.groupby('Customer_ID')['Revenue']  # Groupement par identifiant client
    .sum()                                # Somme des revenus pour chaque client
    .sort_values(ascending=False)         # Tri décroissant (meilleurs clients en premier)
    .reset_index()                        # Réinitialise l'index pour obtenir un DataFrame propre
)

# Création d'un histogramme de distribution des revenus clients
fig = px.histogram(
    customer_revenue,           # DataFrame contenant le CA par client
    x='Revenue',                # Variable à distribuer sur l'axe X
    nbins=50,                   # Nombre de barres (intervalles) pour la distribution
    title='Customer Revenue Distribution'  # Titre du graphique
)

# Application du thème visuel épuré
fig.update_layout(template="plotly_white")

fig.write_image("../reports/figures/Distribution_revenue.png")
# Affichage interactif du graphique
fig.show()

INSIGHT STRATÉGIQUE 

## Principaux enseignements commerciaux

- Le chiffre d'affaires est fortement concentré sur un petit nombre de clients.

- Le Royaume-Uni domine les ventes totales, ce qui indique une forte dépendance à un seul marché.

- Des variations saisonnières marquées suggèrent des opportunités pour des campagnes ciblées.

- Un petit nombre de produits génère la majeure partie du chiffre d'affaires.
